Removed type errors on insert functions

+++ Homework/2/solvers/serial.cpp
@@ -21,7 +21,10 @@ namespace hpc {
 
     auto insert = [&ranges, &bin](/*in*/ const fp value) {
         auto bin_it = std::upper_bound(ranges.cbegin(), ranges.cend(), value);
-        auto index = std::distance(ranges.cbegin(), bin_it);
+        // since bin_it will be always be equal to or ahead of `ranges.begin()`
+        // we can assume that this will always be positive, and therefore no
+        // problem to cast it to a size_t
+        auto index = static_cast<size_t>(std::distance(ranges.cbegin(), bin_it));
 
         bin.counts[index]++;
         if (bin.maxes[index] < value) { bin.maxes[index] = value; }
