+++ Homework/2/solvers/tree.cpp
@@ -116,7 +116,10 @@ namespace detail {
 
         auto insert = [&ranges, &bin](const fp value) {
             auto bin_it = std::upper_bound(ranges.begin(), ranges.end(), value);
-            auto index = std::distance(ranges.begin(), bin_it);
+            // since bin_it will be always be equal to or ahead of `ranges.begin()`
+            // we can assume that this will always be positive, and therefore no
+            // problem to cast it to a size_t
+            auto index = static_cast<size_t>(std::distance(ranges.begin(), bin_it));
 
             bin.counts[index]++;
             if (bin.maxes[index] < value) { bin.maxes[index] = value; }
